Static methods for metrics
